requirements.txt
dominate==2.6.0
imageio
ipython
lpips==0.1.3
matplotlib
numpy>=1.21
opencv-python
pandas
Pillow
scikit-image
tensorboard
tensorboardX
tifffile
torch
torchfile
torchvision
tqdm

